--- BetterThanRottenTomatoAPI/ReviewManagement/Review.Domain/Interfaces/IMovieReviewService.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Review.Domain.Interfaces
-{
-    public interface IMovieReviewService
-    {
-    }
-}
