--- BetterThanRottenTomatoAPI/ReviewManagement/Review.Domain/Services/MovieReviewService.cs
@@ -1,11 +1,0 @@
-﻿using Review.Domain.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Review.Domain.Services
-{
-    public class MovieReviewService :IMovieReviewService
-    {
-    }
-}
