--- BetterThanRottenTomatoAPI/ReviewManagement/Review.Infrastructure/Interfaces/IMovieReviewRepository.cs
@@ -1,14 +1,0 @@
-﻿using Review.Infrastructure.Entities;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Review.Infrastructure.Interfaces
-{
-    public interface IMovieReviewRepository
-    {
-        Task AddMovieReviewByMovieId(long movieId, string comment, int review);
-        Task<Movie> GetMovieReviewsById(long movieId);
-    }
-}
