--- BetterThanRottenTomatoAPI/ReviewManagement/Review.Infrastructure/Repositories/MovieReviewRepository.cs
@@ -1,11 +1,0 @@
-﻿using Review.Infrastructure.Interfaces;
-using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Review.Infrastructure.Repositories
-{
-    public class MovieReviewRepository : IMovieReviewRepository
-    {
-    }
-}
