+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.API/Properties/launchSettings.json
@@ -3,8 +3,8 @@
     "windowsAuthentication": false, 
     "anonymousAuthentication": true, 
     "iisExpress": {
-      "applicationUrl": "http://localhost:49515",
-      "sslPort": 44359
+      "applicationUrl": "http://localhost:61431",
+      "sslPort": 44380
     }
   },
   "profiles": {
