+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.API/ReviewManagement.API.csproj
@@ -1,0 +1,12 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\ReviewManagement.Domain\ReviewManagement.Domain.csproj" />
+    <ProjectReference Include="..\ReviewManagement.Infrastructure\ReviewManagement.Infrastructure.csproj" />
+  </ItemGroup>
+
+</Project>
