+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.API/Startup.cs
@@ -7,8 +7,12 @@ using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
+using ReviewManagement.Domain.Interfaces;
+using ReviewManagement.Domain.Services;
+using ReviewManagement.Infrastructure.Repositories;
+using ReviewManagement.Infrastructure.Repositories.Interfaces;
 
-namespace Review.API
+namespace ReviewManagement.API
 {
     public class Startup
     {
@@ -16,6 +20,12 @@ namespace Review.API
         // For more information on how to configure your application, visit https://go.microsoft.com/fwlink/?LinkID=398940
         public void ConfigureServices(IServiceCollection services)
         {
+            services.AddControllers();
+            services.AddCors();
+
+            services.AddSingleton<IReviewRepository, ReviewRepository>();
+
+            services.AddTransient<IReviewService,ReviewService>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
