+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.API/appsettings.json
@@ -5,5 +5,9 @@
       "Microsoft": "Warning",
       "Microsoft.Hosting.Lifetime": "Information"
     }
+  },
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "TomatoDb": "Server=DESKTOP-LQ9NL1I;Database=BetterThanRottenTomato;Trusted_Connection=True;"
   }
 }
