+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Domain/Services/ReviewService.cs
@@ -1,11 +1,11 @@
-﻿using ReviewManagement.Domain.Services.Interfaces;
+﻿using ReviewManagement.Domain.Interfaces;
 using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace ReviewManagement.Domain.Services
 {
-    public class ReviewService: IReviewService
+    public class ReviewService:IReviewService
     {
     }
 }
