+++ BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/Repositories/Entities/ReviewEntity.cs
@@ -2,11 +2,11 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace ReviewManagement.Infrastructure.Entities
+namespace ReviewManagement.Infrastructure.Repositories.Entities
 {
-    public class Review
+    public class ReviewEntity
     {
-        public long Id { get; set; }
+        public long Id { get; }
         public long MovieId { get; set; }
         public string Username { get; set; }
         public int Rating { get; set; }
