--- BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/ReviewManagement.API/ReviewManagement.API.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-</Project>
