--- BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/ReviewManagement.API/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionString": {
-    "BudgetDb": "Server=DESKTOP-LQ9NL1I;Database=BetterThanRottenTomato;Trusted_Connection=True;"
-  }
-}
