--- BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/ReviewManagement.Domain/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace ReviewManagement.Domain
-{
-    public class Class1
-    {
-    }
-}
