--- BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/ReviewManagement.Domain/ReviewManagement.Domain.csproj
@@ -1,7 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-  </PropertyGroup>
-
-</Project>
