--- BetterThanRottenTomatoAPI/ReviewManagement/ReviewManagement.Infrastructure/ReviewManagement.sln
@@ -1,37 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.29613.14
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReviewManagement.Infrastructure", "ReviewManagement\ReviewManagement.Infrastructure.csproj", "{FAA82B3E-51FC-4546-BFDE-C3C9C2E7D16A}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ReviewManagement.API", "ReviewManagement.API\ReviewManagement.API.csproj", "{0CBF59A2-05C3-4EDA-AA8A-EFEFF2D90F2D}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ReviewManagement.Domain", "ReviewManagement.Domain\ReviewManagement.Domain.csproj", "{5FAB48EC-603D-4B9D-BDEA-B7F3D5066DAC}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{FAA82B3E-51FC-4546-BFDE-C3C9C2E7D16A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{FAA82B3E-51FC-4546-BFDE-C3C9C2E7D16A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{FAA82B3E-51FC-4546-BFDE-C3C9C2E7D16A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{FAA82B3E-51FC-4546-BFDE-C3C9C2E7D16A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{0CBF59A2-05C3-4EDA-AA8A-EFEFF2D90F2D}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0CBF59A2-05C3-4EDA-AA8A-EFEFF2D90F2D}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0CBF59A2-05C3-4EDA-AA8A-EFEFF2D90F2D}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0CBF59A2-05C3-4EDA-AA8A-EFEFF2D90F2D}.Release|Any CPU.Build.0 = Release|Any CPU
-		{5FAB48EC-603D-4B9D-BDEA-B7F3D5066DAC}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{5FAB48EC-603D-4B9D-BDEA-B7F3D5066DAC}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{5FAB48EC-603D-4B9D-BDEA-B7F3D5066DAC}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{5FAB48EC-603D-4B9D-BDEA-B7F3D5066DAC}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {2283CB87-6292-43E6-9279-A40000D44A7E}
-	EndGlobalSection
-EndGlobal
